# src/modules/album/dto/update-album.dto.ts
import { IsNumber, IsOptional, IsString, IsUUID } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class UpdateAlbumDto {
  @ApiProperty({
    description: 'New Album name',
    example: 'Toxicity-V2',
  })
  @IsString()
  name: string;

  @ApiProperty({
    description: 'Album release year',
    example: 2001,
  })
  @IsNumber()
  year: number;

  @ApiProperty({
    description: 'Artist id',
    example: '85a0c8c7-a17e-4c70-8240-e998d2a718e4',
    required: false,
  })
  @IsUUID(4)
  @IsOptional()
  artistId: string;
}